renamed FarmCertificationStepper, move StepIcon to components
